api/requirements.txt
fastapi
openai
uvicorn
pydantic
numpy
tenacity
pytesseract
Pillow
pdf2image
pinecone-client>=2.2.1
PyPDF2>=3.0.0
python-dotenv>=1.0.0
spacy>=3.7.2
python-multipart
easyocr
reportlab
sqlmodel
pypdfium2
chroma-hnswlib>=0.7.0